Added WIP distance to boundary
Distance to boundary is to add adding Lift to a particle in flow. It will look up an array of vertices to compare the current particle position to. Currently this is in Python and is incredibly slow. Looking at using cpp method but requires understanding of boundary domains in LEoPart.