Fixed issues Lift and dT
Lift had an issue which would ignore the particle position relative to the boundary as well as investigating automating the time step, dT, using ANSYS discrete phase modelling (DPM) a.k.a Lagrangian Particle Tracking (LPT). The method aims to use a particle relaxation factor divided by 5. If the dT is larger, it is replaced with the particle relaxation factor divided by 5.